Move function inRange to ChartUtils (#4556)
