+++ src/chart/generateCategoricalChart.tsx
@@ -42,6 +42,7 @@ import {
   getStackGroupsByAxisId,
   getTicksOfAxis,
   getTooltipItem,
+  inRange,
   isAxisLTR,
   isCategoricalAxis,
   parseDomainOfCategoryAxis,
@@ -49,7 +50,7 @@ import {
   parseSpecifiedDomain,
 } from '../util/ChartUtils';
 import { detectReferenceElementsDomain } from '../util/DetectReferenceElementsDomain';
-import { inRangeOfSector, polarToCartesian } from '../util/PolarUtils';
+import { polarToCartesian } from '../util/PolarUtils';
 import { shallowEqual } from '../util/ShallowEqual';
 import { eventCenter, SYNC_EVENT } from '../util/Events';
 import {
@@ -213,7 +214,7 @@ function getDefaultDomainByAxisType(axisType: 'number' | string) {
 }
 
 /**
- * Get the content to be displayed in the tooltip
+ * @deprecated this indirectly depends on the list of all children read from DOM. Use Redux instead.
  * @param  {Object} state          Current state
  * @param  {Array}  chartData      The data defined in chart
  * @param  {Number} activeIndex    Active index of data
@@ -1367,7 +1368,15 @@ export const generateCategoricalChart = ({
 
       const scale = boundingRect.width / element.offsetWidth || 1;
 
-      const rangeObj = this.inRange(e.chartX, e.chartY, scale);
+      const rangeObj = inRange(
+        e.chartX,
+        e.chartY,
+        scale,
+        this.props.layout,
+        this.state.angleAxisMap,
+        this.state.radiusAxisMap,
+        this.state.offset,
+      );
       if (!rangeObj) {
         return null;
       }
@@ -1391,33 +1400,6 @@ export const generateCategoricalChart = ({
           ...e,
           ...toolTipData,
         };
-      }
-
-      return null;
-    }
-
-    inRange(x: number, y: number, scale = 1): RangeObj {
-      const { layout } = this.props;
-
-      const [scaledX, scaledY] = [x / scale, y / scale];
-
-      if (layout === 'horizontal' || layout === 'vertical') {
-        const { offset } = this.state;
-
-        const isInRange =
-          scaledX >= offset.left &&
-          scaledX <= offset.left + offset.width &&
-          scaledY >= offset.top &&
-          scaledY <= offset.top + offset.height;
-
-        return isInRange ? { x: scaledX, y: scaledY } : null;
-      }
-
-      const { angleAxisMap, radiusAxisMap } = this.state;
-
-      if (angleAxisMap && radiusAxisMap) {
-        const angleAxis = getAnyElementOfObject(angleAxisMap);
-        return inRangeOfSector({ x: scaledX, y: scaledY }, angleAxis);
       }
 
       return null;
